Simulate MRS data with a similar appearance to normal brain (by default).

Simulate MRS data with a similar appearance to normal brain (by default).

sim_brain_1h(acq_paras = def_acq_paras(), type = "normal",
  pul_seq = seq_press_ideal, xlim = c(0.5, 4.2), full_output = FALSE,
  amps = NULL, ...)

Arguments

acq_paras
list of acquisition parameters or an mrs_data object. See def_acq_paras

type
type of spectrum, only "normal" is implemented currently.

pul_seq
pulse sequence function to use

xlim
range of frequencies to simulate in ppm

full_output
when FALSE (default) only output the simulated MRS data. When TRUE output a list containing the MRS data, basis set object and corresponding amplitudes.

amps
a vector of basis amplitudes may be specified to modify the output spectrum

...
extra parameters to pass to the pulse sequence function.

Value

see full_output option