Return a list of options for VARPRO based fitting with 3 free parameters:

zero'th order phase correction

global damping

global frequency shift.

Return a list of options for VARPRO based fitting with 3 free parameters:

zero'th order phase correction

global damping

global frequency shift.

varpro_3_para_opts(nstart = 20, init_damping = 2, maxiters = 200,
  max_shift = 5, max_damping = 5, anal_jac = FALSE, bl_smth_pts = 80)

Arguments

nstart
Position in the time-domain to start fitting, units of data points.

init_damping
Starting value for the global Gaussian line-broadening term - measured in Hz.

maxiters
Maximum number of levmar iterations to perform.

max_shift
Maximum global shift allowed, measured in Hz.

max_damping
Maximum damping allowed, FWHM measured in Hz.

anal_jac
Option to use the analytic or numerical Jacobian (logical).

bl_smth_pts
Number of data points to use in the baseline smoothing calculation.

Value

List of options.